Fix CredRank JSON serialization
As a prelude to node virtualization, I've added a test to verify that
JSON serialization is a proper round trip. In so doing, I discovered
that our current JSON serialization does not work as intended, due to a
subtle issue where we store markov edges by their address property,
which (as the docs note) is not actually a full address. I'm considering
that we may want to refactor so that `MarkovEdgeAddress.address` is
actually a markov edge address, to avoid confusion like this in the
future.

Test plan: `yarn test`, passes, and credrank scores are stable.