Virtualize the fibration edges
This commit marks the first batch of edge virtualization. I decided to
virtualize all of the fibration edges (payout, forward webbing, and
backward webbing) in one step since they are all so similar.

Now this is space saving! We're down to 57% of original.

| Step | Virtualize | # Nodes | # Edges | Disk Space | % of full |
| ---- | ---------- | ------- | ------- | ---------- | -------- |
| 0 | None | 96144 | 416365 | 116M | 100% |
| 1 | Seed | 96143 | 416365 | 116M | 100% |
| 2 | Accumulators | 96001 | 416365 | 116M | 100% |
| 3 | Epochs | 30823 | 416365 | 102M | 88% |
| 4 | Fibrations | 30823 | 220833 | 66M | 57% |

Warning: When running this variant, I sense a slight disturbance in the
Cred. We should track it down and understand what's causing it.